requirements.txt
python-dotenv
requests
redis